user_dashboard: define route handlers in each file
