+++ backend/ct_dashboard_api/src/routes/v1/index.ts
@@ -16,14 +16,10 @@ import { customerJwtMiddleware } from "@oko-wallet-ctd-api/middleware/auth";
 import { customerLogoUploadMiddleware } from "@oko-wallet-ctd-api/middleware/multer";
 import { rateLimitMiddleware } from "@oko-wallet-ctd-api/middleware/rate_limit";
 
-const ONE_MINUTE_SEC = 1 * 60;
-
 export function makeCustomerRouter() {
   const router = express.Router();
 
-  router.use(
-    rateLimitMiddleware({ windowSeconds: ONE_MINUTE_SEC, maxRequests: 30 }),
-  );
+  router.use(rateLimitMiddleware({ windowSeconds: 60, maxRequests: 30 }));
 
   router.post(
     "/customer/auth/forgot-password",
