+++ backend/user_dashboard_api/src/routes/user_auth.test.ts
@@ -68,6 +68,7 @@ describe("Customer Auth API Integration Tests", () => {
       status: "ACTIVE",
       url: "",
       logo_url: "",
+      theme: "system",
     };
 
     testCustomerDashboardUser = {
