+++ backend/user_dashboard_api/src/routes/verify_email_and_login.ts
@@ -1,0 +1,230 @@
+import { comparePassword, hashPassword } from "@oko-wallet/crypto-js";
+import { ErrorCodeMap } from "@oko-wallet/oko-api-error-codes";
+import { registry } from "@oko-wallet/oko-api-openapi";
+import { ErrorResponseSchema } from "@oko-wallet/oko-api-openapi/common";
+import {
+  ChangePasswordRequestSchema,
+  ChangePasswordSuccessResponseSchema,
+  CustomerAuthHeaderSchema,
+  LoginSuccessResponseSchema,
+  SendVerificationRequestSchema,
+  SendVerificationSuccessResponseSchema,
+  SignInRequestSchema,
+  VerifyAndLoginRequestSchema,
+} from "@oko-wallet/oko-api-openapi/ct_dashboard";
+import {
+  getCTDUserWithCustomerAndPasswordHashByEmail,
+  getCTDUserWithCustomerByEmail,
+  updateCustomerDashboardUserPassword,
+  verifyCustomerDashboardUserEmail,
+} from "@oko-wallet/oko-pg-interface/customer_dashboard_users";
+import { verifyEmailCode } from "@oko-wallet/oko-pg-interface/email_verifications";
+import type { OkoApiResponse } from "@oko-wallet/oko-types/api_response";
+import type {
+  ChangePasswordRequest,
+  ChangePasswordResponse,
+  LoginResponse,
+  SendVerificationRequest,
+  SendVerificationResponse,
+  SignInRequest,
+  VerifyAndLoginRequest,
+} from "@oko-wallet/oko-types/ct_dashboard";
+import type { Request, Response, Router } from "express";
+
+import { sendVerificationCodeRoute } from "./send_verification_code";
+import { generateCustomerToken } from "@oko-wallet-usrd-api/auth";
+import {
+  CHANGED_PASSWORD_MIN_LENGTH,
+  EMAIL_REGEX,
+  SIX_DIGITS_REGEX,
+} from "@oko-wallet-usrd-api/constants";
+import { sendEmailVerificationCode } from "@oko-wallet-usrd-api/email/send";
+import {
+  type CustomerAuthenticatedRequest,
+  customerJwtMiddleware,
+} from "@oko-wallet-usrd-api/middleware/auth";
+import { rateLimitMiddleware } from "@oko-wallet-usrd-api/middleware/rate_limit";
+
+registry.registerPath({
+  method: "post",
+  path: "/customer_dashboard/v1/customer/auth/verify-login",
+  tags: ["Customer Dashboard"],
+  summary: "Verify email and login",
+  description:
+    "Verifies the email with provided code and logs in the customer dashboard user",
+  security: [],
+  request: {
+    body: {
+      required: true,
+      content: {
+        "application/json": {
+          schema: VerifyAndLoginRequestSchema,
+        },
+      },
+    },
+  },
+  responses: {
+    200: {
+      description: "Successfully verified and logged in",
+      content: {
+        "application/json": {
+          schema: LoginSuccessResponseSchema,
+        },
+      },
+    },
+    400: {
+      description: "Invalid request",
+      content: {
+        "application/json": {
+          schema: ErrorResponseSchema,
+        },
+      },
+    },
+    404: {
+      description: "Account not found",
+      content: {
+        "application/json": {
+          schema: ErrorResponseSchema,
+        },
+      },
+    },
+    500: {
+      description: "Server error",
+      content: {
+        "application/json": {
+          schema: ErrorResponseSchema,
+        },
+      },
+    },
+  },
+});
+
+export async function verifyEmailAndLogin(
+  req: Request,
+  res: Response<OkoApiResponse<LoginResponse>>,
+) {
+  try {
+    const state = req.app.locals as any;
+    const request: VerifyAndLoginRequest = req.body;
+
+    if (!request.email || !request.verification_code) {
+      res.status(400).json({
+        success: false,
+        code: "CUSTOMER_ACCOUNT_NOT_FOUND",
+        msg: "email and verification_code are required",
+      });
+      return;
+    }
+
+    // Validate verification code format (6 digits)
+    if (!SIX_DIGITS_REGEX.test(request.verification_code)) {
+      res.status(400).json({
+        success: false,
+        code: "INVALID_VERIFICATION_CODE",
+        msg: "Verification code must be 6 digits",
+      });
+      return;
+    }
+
+    // Inline verifyCodeAndLogin logic
+    const customerAccountResult = await getCTDUserWithCustomerByEmail(
+      state.db,
+      request.email,
+    );
+    if (!customerAccountResult.success) {
+      res.status(500).json({
+        success: false,
+        code: "UNKNOWN_ERROR",
+        msg: `Failed to get customer account: ${customerAccountResult.err}`,
+      });
+      return;
+    }
+    const customerAccount = customerAccountResult.data;
+
+    if (customerAccount === null) {
+      res.status(404).json({
+        success: false,
+        code: "CUSTOMER_ACCOUNT_NOT_FOUND",
+        msg: "Account not found",
+      });
+      return;
+    }
+
+    if (customerAccount.user.is_email_verified) {
+      res.status(400).json({
+        success: false,
+        code: "EMAIL_ALREADY_VERIFIED",
+        msg: "Email already verified",
+      });
+      return;
+    }
+
+    // Verify the code first
+    const verificationResult = await verifyEmailCode(state.db, {
+      email: request.email,
+      verification_code: request.verification_code,
+    });
+
+    if (!verificationResult.success) {
+      res.status(400).json({
+        success: false,
+        code: "INVALID_VERIFICATION_CODE",
+        msg: `Invalid or expired verification code: ${verificationResult.err}`,
+      });
+      return;
+    }
+
+    const verifyCustomerAccountEmailResult =
+      await verifyCustomerDashboardUserEmail(state.db, {
+        user_id: customerAccount.user.user_id,
+      });
+
+    if (!verifyCustomerAccountEmailResult.success) {
+      res.status(500).json({
+        success: false,
+        code: "UNKNOWN_ERROR",
+        msg: `Failed to verify email: ${verifyCustomerAccountEmailResult.err}`,
+      });
+      return;
+    }
+
+    // Generate JWT token
+    const tokenResult = generateCustomerToken({
+      user_id: customerAccount.user.user_id,
+      jwt_config: {
+        secret: state.jwt_secret,
+        expires_in: state.jwt_expires_in,
+      },
+    });
+
+    if (!tokenResult.success) {
+      res.status(500).json({
+        success: false,
+        code: "FAILED_TO_GENERATE_TOKEN",
+        msg: `Failed to generate authentication token: ${tokenResult.err}`,
+      });
+      return;
+    }
+
+    res.status(200).json({
+      success: true,
+      data: {
+        token: tokenResult.data.token,
+        customer: {
+          email: request.email,
+          is_email_verified:
+            verifyCustomerAccountEmailResult.data.is_email_verified,
+        },
+      },
+    });
+    return;
+  } catch (error) {
+    console.error("Verify and login route error:", error);
+    res.status(500).json({
+      success: false,
+      code: "UNKNOWN_ERROR",
+      msg: `Internal server error: ${error instanceof Error ? error.message : String(error)}`,
+    });
+    return;
+  }
+}
